appium-tests/requirements.txt
Appium-Python-Client
pytest
browserstack-sdk
Pillow
